LibraryAPI/requirements.txt
-i https://pypi.org/simple
alembic==1.13.2; python_version >= '3.8'
blinker==1.8.2; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation == 'CPython' and sys_platform == 'win32'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
flask==3.0.3; python_version >= '3.8'
flask-marshmallow==1.2.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.5; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.22.0; python_version >= '3.8'
marshmallow-sqlalchemy==1.1.0; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
setuptools==75.1.0; python_version >= '3.8'
sqlalchemy==2.0.35; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
werkzeug==3.0.4; python_version >= '3.8'
zope.event==5.0; python_version >= '3.7'
zope.interface==7.0.3; python_version >= '3.8'
